Add some lints check and fix some errors

+++ fuse_ll/src/fuse/channel.rs
@@ -91,7 +91,7 @@ impl Channel {
         match res {
             Ok(s) => {
                 unsafe {
-                    buffer.set_len(s as usize);
+                    buffer.set_len(s);
                 }
                 debug!("receive successfully {} byte data", s);
                 Ok(())
